src/api/requirements.txt
fastapi
fastapi[standard]
requests
gunicorn==21.2.0
azure-cosmos
azure-search-documents==11.4.0
azure-keyvault-secrets
aiohttp==3.9.5
python-dotenv==1.0.1
azure-cognitiveservices-speech==1.37.0
prompty[azure]==0.1.24
jsonlines
nbconvert
pandas
opentelemetry-instrumentation-fastapi
jupyter
opentelemetry-instrumentation
azure-identity==1.17.1
tabulate
azure-monitor-events-extension
azure-appconfiguration-provider==2.0.0b2
FeatureManagement==2.0.0b2
azure-monitor-opentelemetry
azure-monitor-opentelemetry-exporter
azure-identity